+++ lib/instabug_flutter.dart
@@ -1,7 +1,7 @@
 // Models
 export 'src/models/crash_data.dart';
 export 'src/models/exception_data.dart';
-export 'src/models/ibg_feature_flag.dart';
+export 'src/models/feature_flag.dart';
 export 'src/models/network_data.dart';
 export 'src/models/trace.dart';
 // Modules